+++ bin/javadoc.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
 
-exec mvn javadoc:javadoc
+#exec mvn javadoc:javadoc
+
+# 多模块聚合文档，生成在target/site/apidocs
+exec mvn javadoc:aggregate