+++ hutool-core/src/main/java/cn/hutool/core/util/TypeUtil.java
@@ -1,6 +1,5 @@
 package cn.hutool.core.util;
 
-import cn.hutool.core.collection.ListUtil;
 import cn.hutool.core.lang.ParameterizedTypeImpl;
 import cn.hutool.core.lang.reflect.ActualTypeMapperPool;
 
@@ -41,7 +40,10 @@ public class TypeUtil {
 			} else if (type instanceof ParameterizedType) {
 				return (Class<?>) ((ParameterizedType) type).getRawType();
 			} else if (type instanceof TypeVariable) {
-				return (Class<?>) ((TypeVariable<?>) type).getBounds()[0];
+				Type[] bounds = ((TypeVariable<?>) type).getBounds();
+				if (bounds.length == 1) {
+					return getClass(bounds[0]);
+				}
 			} else if (type instanceof WildcardType) {
 				final Type[] upperBounds = ((WildcardType) type).getUpperBounds();
 				if (upperBounds.length == 1) {