+++ hutool-cron/src/main/java/cn/hutool/cron/pattern/parser/PartParser.java
@@ -3,16 +3,11 @@ package cn.hutool.cron.pattern.parser;
 import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.date.Month;
 import cn.hutool.core.date.Week;
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.util.NumberUtil;
 import cn.hutool.core.util.StrUtil;
 import cn.hutool.cron.CronException;
 import cn.hutool.cron.pattern.Part;
-import cn.hutool.cron.pattern.matcher.AlwaysTrueMatcher;
-import cn.hutool.cron.pattern.matcher.BoolArrayMatcher;
-import cn.hutool.cron.pattern.matcher.DayOfMonthMatcher;
-import cn.hutool.cron.pattern.matcher.PartMatcher;
-import cn.hutool.cron.pattern.matcher.YearValueMatcher;
+import cn.hutool.cron.pattern.matcher.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -204,13 +199,11 @@ public class PartParser {
 				//在range模式下，如果步进不存在，表示步进为1
 				step = 1;
 			}
-			if (v1 < v2) {// 正常范围，例如：2-5
+			if (v1 <= v2) {// 正常范围，例如：2-5，3-3
 				NumberUtil.appendRange(v1, v2, step, results);
-			} else if (v1 > v2) {// 逆向范围，反选模式，例如：5-2
+			} else {// 逆向范围，反选模式，例如：5-2
 				NumberUtil.appendRange(v1, part.getMax(), step, results);
 				NumberUtil.appendRange(part.getMin(), v2, step, results);
-			} else {// v1 == v2，此时与单值模式一致
-				NumberUtil.appendRange(v1, part.getMax(), step, results);
 			}
 		} else {
 			throw new CronException("Invalid syntax of field: [{}]", value);