+++ hutool-json/src/main/java/cn/hutool/json/JSONParser.java
@@ -1,6 +1,5 @@
 package cn.hutool.json;
 
-import cn.hutool.core.lang.Console;
 import cn.hutool.core.lang.Filter;
 import cn.hutool.core.lang.mutable.Mutable;
 import cn.hutool.core.lang.mutable.MutablePair;