+++ hutool-json/src/test/java/cn/hutool/json/IssueI6YN2ATest.java
@@ -12,8 +12,6 @@
 
 package cn.hutool.json;
 
-import cn.hutool.core.lang.Console;
-import cn.hutool.core.lang.TypeReference;
 import lombok.Data;
 import org.junit.Assert;
 import org.junit.Test;