+++ hutool-jwt/src/test/java/cn/hutool/jwt/Issue3205Test.java
@@ -30,7 +30,7 @@ public class Issue3205Test {
 
 		final String token = jwt.sign();
 
-		final boolean signed = Jwts.parserBuilder().setSigningKey(keyPair.getPublic()).build().isSigned(token);
+		final boolean signed = Jwts.parser().verifyWith(keyPair.getPublic()).build().isSigned(token);
 
 		Assert.assertTrue(signed);
 	}